polling hook wrapper
